Tell Emscripten not to minify the HTML shell file since it causes errors
For some reason Emscripten's HTML minification step introduces errors into the
markup, which is very sloppy. Examples are missing the space in the doctype
declaration and messing with meta tags in some ways.

Since our HTML shell file is tiny we will just not bother minifying it. If it
was large we would want to introduce a separate and better minification step.